fix numeric order, refactor Document
fix numeric order to prefer i64
rename and move Document stuff
